Added lots of stuff
I don't even remember half of what I did, I'm tired, like really tired....

+++ Assets/Scripts/Inventory.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public static class Inventory
 {
-    public static short ammo = 100;
-    public static short laserEnergy = 100;
+    public static int ammo = 100;
+    public static int laserEnergy = 100;
 }
